pyproject.toml
[project]
name = "background-remover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onnxruntime>=1.22.0",
    "rembg>=2.0.66",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pillow>=11.0.0",
]
